+++ changelog.md
@@ -1,4 +1,12 @@
 # Changelog
 
-**./public/index.html**
-* Updated page text to refer to the FOX Controller.
+**./fox-api/alarm-database.js**
+* Renamed 'getAlarmObjects' to 'getPopulatedAlarmObjects'
+	* Called publicly as 'getPopulatedAlarms'
+	* Uncommented `queryRes.push`
+	* Returns array of alarm objects that match input parameters.
+* Wrote new function 'getPopulatedAvailabilityObjects'
+	* GET function for 'available' global.
+	* Returns array of alarm availability objects.
+	* Called publicly as 'getPopulatedAvailable'
+* Rewrote 'getAlarmObjects' to only return an empty array.